# swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /shop:
    get:
      operationId: AppController_getShop
      parameters:
        - name: page
          required: true
          in: query
          schema:
            type: number
        - name: limit
          required: true
          in: query
          schema:
            type: number
      responses:
        "200":
          description: ""
...
  /photo/create:
    post:
      operationId: CommentController_create
      summary: Create a new comment
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCommentBody"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      tags: &a1
        - comments
      security: &a2
        - bearer: []
  /photo:
    get:
      operationId: CommentController_findAll
      summary: Get all comments
      parameters: []
      responses:
        "200":
          description: Comments found.
        "404":
          description: No comments found.
      tags: *a1
      security: *a2
  /photo/{id}:
    get:
      operationId: CommentController_getComment
      summary: Get a comment by id
      parameters:
        - name: id
          required: true
          in: path
          description: The id of the comment
          schema:
            type: number
      responses:
        "200":
          description: Comment found.
        "404":
          description: Comment not found.
      tags: *a1
      security: *a2
    delete:
      operationId: CommentController_deleteComment
      summary: Delete a comment
      parameters:
        - name: id
          required: true
          in: path
          description: The id of the comment to delete
          schema:
            type: number
      responses:
        "200":
          description: Comment deleted.
        "404":
          description: Comment not found.
      tags: *a1
      security: *a2
    put:
      operationId: CommentController_updateComment
      summary: Update a comment by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "200":
          description: Comment successfully updated.
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      tags: *a1
      security: *a2
...
  /attraction/creat:
    post:
      operationId: AttractionController_create
      summary: Create a new attraction
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttractionBody"
      responses:
        "201":
          description: Attraction successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAttractionBody"
        "400":
          description: Bad Request
      tags: &a7
        - attractions
  /attraction:
    get:
      operationId: AttractionController_findAll
      summary: Get all attractions
      parameters: []
      responses:
        "200":
          description: Attractions found.
        "404":
          description: No attractions found.
      tags: *a7
  /attraction/{id}:
    get:
      operationId: AttractionController_getAttraction
      summary: Get an attraction by id
      parameters:
        - name: id
          required: true
          in: path
          description: The id of the attraction
          schema:
            type: number
      responses:
        "200":
          description: Attraction found.
        "404":
          description: Attraction not found.
      tags: *a7
    delete:
      operationId: AttractionController_deleteAttraction
      summary: Delete an attraction
      parameters:
        - name: id
          required: true
          in: path
          description: The id of the attraction to delete
          schema:
            type: number
      responses:
        "200":
          description: Attraction deleted.
        "404":
          description: Attraction not found.
      tags: *a7
    put:
      operationId: AttractionController_updateAttraction
      summary: Update an attraction by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttractionBody"
      responses:
        "200":
          description: Attraction successfully updated.
        "400":
          description: Bad Request
      tags: *a7
  /cart/add:
    post:
      operationId: CartController_addToCart
      summary: Add product to cart
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatCartBody"
      responses:
        "201":
          description: Product successfully added to cart.
        "400":
          description: Bad Request
      tags: &a8
        - cart
      security: &a9
        - bearer: []
  /cart/{userId}:
    get:
      operationId: CartController_getCart
      summary: Get cart by user ID
      parameters:
        - name: userId
          required: true
          in: path
          description: User ID
          schema:
            type: number
      responses:
        "200":
          description: Cart found.
        "404":
          description: Cart not found.
      tags: *a8
      security: *a9
  /cart/by-id/{cartId}:
    get:
      operationId: CartController_getCartById
      summary: Get cart by cart ID
      parameters:
        - name: cartId
          required: true
          in: path
          description: Cart ID
          schema:
            type: number
      responses:
        "200":
          description: Cart found.
        "404":
          description: Cart not found.
      tags: *a8
      security: *a9
  /cart/item/{cartItemId}:
    put:
      operationId: CartController_updateCartItem
      summary: Update cart item quantity
      parameters:
        - name: cartItemId
          required: true
          in: path
          description: Cart Item ID
          schema:
            type: number
      responses:
        "200":
          description: Quantity updated.
        "404":
          description: Cart item not found.
      tags: *a8
      security: *a9
    delete:
      operationId: CartController_removeCartItem
      summary: Remove item from cart
      parameters:
        - name: cartItemId
          required: true
          in: path
          description: Cart Item ID
          schema:
            type: number
      responses:
        "200":
          description: Item removed.
        "404":
          description: Item not found.
      tags: *a8
      security: *a9
info:
  title: Salekhard
  description: Welcome to Salekhard
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    basic:
      type: http
      scheme: basic
  schemas:
    CreateCommentBody:
      type: object
      properties:
        name:
          type: string
          example: Sergey
          description: The name of the commenter
        email:
          type: string
          example: sergey@example.com
          description: The email address of the commenter
        content:
          type: string
          example: Great attraction!
          description: The content of the comment
      required:
        - name
        - email
        - content
...
    CreateAttractionBody:
      type: object
      properties:
        name:
          type: string
          example: Name of the Attraction
          description: The name of the attraction
        location:
          type: string
          example: Location of the Attraction
          description: The location of the attraction
        foundationYear:
          type: number
          example: 1990
          description: The year the attraction was founded
        description:
          type: string
          example: Description of the Attraction
          description: The description of the attraction
        image:
          type: string
          example: /images/attraction.jpg
          description: The image path of the attraction
      required:
        - name
        - location
        - foundationYear
        - description
        - image
    CreatCartBody:
      type: object
      properties:
        userId:
          type: number
          example: 1
          description: The ID of the user
        productId:
          type: number
          example: 1
          description: The ID of the product to add to the cart
        quantity:
          type: number
          example: 2
          description: Quantity of the product
      required:
        - userId
        - productId
        - quantity